Ripping off the bandaid by merging files
