+++ case_studies/sdss_galaxies/Makefile
@@ -1,4 +1,4 @@
-all: models/sdss_autoencoder.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_sleep.pt
+all: models/sdss_autoencoder.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_location_encoder.pt
 
 models/sdss_autoencoder.pt:
 	./run_experiment.sh sdss_autoencoder
@@ -10,11 +10,11 @@ models/sdss_binary.pt: models/sdss_autoencoder.pt
 models/sdss_galaxy_encoder.pt: models/sdss_autoencoder.pt
 	./run_experiment.sh sdss_galaxy_encoder
 
-models/sdss_sleep.pt: models/sdss_autoencoder.pt
-	./run_experiment.sh sdss_sleep
+models/sdss_location_encoder.pt: models/sdss_autoencoder.pt
+	./run_experiment.sh sdss_location_encoder
 
-models/sdss_galaxy_encoder_real.pt: models/sdss_sleep.pt models/sdss_autoencoder.pt
+models/sdss_galaxy_encoder_real.pt: models/sdss_location_encoder.pt models/sdss_autoencoder.pt
 	./run_experiment.sh sdss_galaxy_encoder_real
 
 clean:
-	rm models/sdss_autoencoder.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_sleep.pt models/latents_simulated_sdss_galaxies.pt
+	rm models/sdss_autoencoder.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_location_encoder.pt models/latents_simulated_sdss_galaxies.pt
