+++ case_studies/sdss_galaxies/config/config.yaml
@@ -114,7 +114,7 @@ models:
         psf_slen: 25
         sdss_bands:
             - 2
-    encoder:
+    location_encoder:
         _target_: bliss.models.location_encoder.LocationEncoder
         input_transform:
             _target_: bliss.models.location_encoder.ConcatBackgroundTransform
@@ -127,6 +127,8 @@ models:
         spatial_dropout: 0.0
         dropout: 0.0
         hidden: 128
+        annotate_probs: True
+        slack: 1.0
     galaxy_encoder:
         _target_: bliss.models.galaxy_encoder.GalaxyEncoder
         decoder: ${models.decoder}
@@ -156,12 +158,6 @@ models:
             latents_file: ${paths.project}/models/latents_simulated_sdss_galaxies.pt
             psf_image_file: ${paths.data}/psField-000094-1-0012-PSF-image.npy
             galaxy_dataset: ${datasets.sdss_galaxies}
-    sleep:
-        _target_: bliss.sleep.SleepPhase
-        encoder: ${models.encoder}
-        decoder: ${models.decoder}
-        annotate_probs: True
-        slack: 1.0
 
 generate:
     dataset:
