+++ case_studies/sdss_galaxies_vae/Makefile
@@ -3,7 +3,7 @@ GPU_BINARY:=$(GPU)
 GPU_GALAXY_ENCODER:=$(GPU)
 GPU_SLEEP:=$(GPU)
 
-all: models/sdss_vae.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_sleep.pt models/sdss_galaxy_encoder_real.pt
+all: models/sdss_vae.pt models/sdss_binary.pt models/sdss_galaxy_encoder.pt models/sdss_location_encoder.pt models/sdss_galaxy_encoder_real.pt
 
 models/sdss_vae.pt:
 	./run_experiment.sh sdss_vae $(GPU)
@@ -17,8 +17,8 @@ models/sdss_binary.pt: models/sdss_vae.pt
 models/sdss_galaxy_encoder.pt: models/sdss_vae.pt
 	./run_experiment.sh sdss_galaxy_encoder $(GPU)
 
-models/sdss_sleep.pt: models/sdss_vae.pt
-	./run_experiment.sh sdss_sleep $(GPU_SLEEP)
+models/sdss_location_encoder.pt: models/sdss_vae.pt
+	./run_experiment.sh sdss_location_encoder $(GPU_SLEEP)
 
-models/sdss_galaxy_encoder_real.pt: models/sdss_sleep.pt models/sdss_binary.pt models/sdss_vae.pt models/sdss_galaxy_encoder.pt
+models/sdss_galaxy_encoder_real.pt: models/sdss_location_encoder.pt models/sdss_binary.pt models/sdss_vae.pt models/sdss_galaxy_encoder.pt
 	./run_experiment.sh sdss_galaxy_encoder_real $(GPU_SLEEP)
