+++ case_studies/sdss_galaxies_vae/config/training/sdss_location_encoder.yaml
@@ -1,7 +1,7 @@
 # @package _global_
 training:
-  name: "sdss_sleep"
-  model: ${models.sleep}
+  name: "sdss_location_encoder"
+  model: ${models.location_encoder}
   dataset: ${datasets.simulated}
   optimizer_params:
     lr: 1e-4
